+++ src/StreamController.cpp
@@ -6,6 +6,7 @@
 #include "Stream/Partition.h"
 #include "TimeUtility.h"
 #include "helper.h"
+#include "logger.h"
 
 namespace FileWriter {
 StreamController::StreamController(
@@ -17,7 +18,7 @@ StreamController::StreamController(
       WriterThread([this]() { WriterTask->flushDataToFile(); },
                    Settings.DataFlushInterval,
                    Registrar.getNewRegistrar("stream")),
-      KafkaSettings(Settings), MetaDataTracker(Tracker) {
+      StreamerOptions(Settings), MetaDataTracker(Tracker) {
   Executor.sendLowPriorityWork([=]() {
     CurrentMetadataTimeOut = Settings.BrokerSettings.MinMetadataTimeout;
     getTopicNames();
@@ -32,11 +33,31 @@ StreamController::~StreamController() {
 
 void StreamController::setStopTime(time_point const &StopTime) {
   Executor.sendWork([=]() {
-    KafkaSettings.StopTimestamp = StopTime;
+    StreamerOptions.StopTimestamp = StopTime;
     for (auto &s : Streamers) {
       s->setStopTime(StopTime);
     }
   });
+}
+
+void StreamController::pauseStreamers() {
+  for (auto &Stream : Streamers) {
+    Stream->pause();
+  }
+  // We set the mutex last. Pause operations are idempotent, so we worry more
+  // about preventing a concurrent call to resumeStreamers() than a concurrent
+  // call to pauseStreamers()
+  StreamersPaused.store(true);
+}
+
+void StreamController::resumeStreamers() {
+  for (auto &Stream : Streamers) {
+    Stream->resume();
+  }
+  // We set the mutex last. Resume operations are idempotent, so we worry more
+  // about preventing a concurrent call to pauseStreamers() than a concurrent
+  // call to resumeStreamers()
+  StreamersPaused.store(false);
 }
 
 void StreamController::stop() {
@@ -53,8 +74,8 @@ bool StreamController::isDoneWriting() {
   auto IsDoneWriting =
       HasError or StopNow or
       (!StreamersRemaining.load() and
-       KafkaSettings.StopTimestamp != time_point(duration(0)) and
-       Now > KafkaSettings.StopTimestamp);
+       StreamerOptions.StopTimestamp != time_point(duration(0)) and
+       Now > StreamerOptions.StopTimestamp);
   if (not IsDoneWriting) {
     auto TimeDiffPeriods = (Now - LastFileSizeCalcTime) / FileSizeCalcInterval;
     if (TimeDiffPeriods >= 1) {
@@ -70,13 +91,13 @@ std::string StreamController::getJobId() const { return WriterTask->jobID(); }
 
 void StreamController::getTopicNames() {
   try {
-    auto TopicNames = Kafka::getTopicList(KafkaSettings.BrokerSettings.Address,
-                                          CurrentMetadataTimeOut,
-                                          KafkaSettings.BrokerSettings);
+    auto TopicNames = Kafka::getTopicList(
+        StreamerOptions.BrokerSettings.Address, CurrentMetadataTimeOut,
+        StreamerOptions.BrokerSettings);
     Executor.sendLowPriorityWork([=]() { initStreams(TopicNames); });
   } catch (MetadataException &E) {
     CurrentMetadataTimeOut *= 2;
-    auto &Settings = KafkaSettings.BrokerSettings;
+    auto &Settings = StreamerOptions.BrokerSettings;
     if (CurrentMetadataTimeOut > Settings.MaxMetadataTimeout) {
       CurrentMetadataTimeOut = Settings.MaxMetadataTimeout;
     }
@@ -114,17 +135,18 @@ void StreamController::initStreams(std::set<std::string> KnownTopicNames) {
   }
   for (auto &CItem : TopicSrcMap) {
     auto CStartTime =
-        std::chrono::system_clock::time_point(KafkaSettings.StartTimestamp);
+        std::chrono::system_clock::time_point(StreamerOptions.StartTimestamp);
     auto CStopTime =
-        std::chrono::system_clock::time_point(KafkaSettings.StopTimestamp);
+        std::chrono::system_clock::time_point(StreamerOptions.StopTimestamp);
     auto CTopic = std::make_unique<Stream::Topic>(
-        KafkaSettings.BrokerSettings, CItem.first, CItem.second, &WriterThread,
-        StreamMetricRegistrar, CStartTime, KafkaSettings.BeforeStartTime,
-        CStopTime, KafkaSettings.AfterStopTime);
+        StreamerOptions.BrokerSettings, CItem.first, CItem.second,
+        &WriterThread, StreamMetricRegistrar, CStartTime,
+        StreamerOptions.BeforeStartTime, CStopTime,
+        StreamerOptions.AfterStopTime);
     CTopic->start();
     Streamers.emplace_back(std::move(CTopic));
   }
-  Executor.sendLowPriorityWork([=]() { checkIfStreamsAreDone(); });
+  Executor.sendLowPriorityWork([=]() { performPeriodicChecks(); });
 }
 
 bool StreamController::hasErrorState() const { return HasError; }
@@ -134,19 +156,22 @@ std::string StreamController::errorMessage() {
   return ErrorMessage;
 }
 
-using std::chrono_literals::operator""ms;
+void StreamController::performPeriodicChecks() {
+  checkIfStreamsAreDone();
+  throttleIfWriteQueueIsFull();
+  std::this_thread::sleep_for(PeriodicChecksInterval);
+  Executor.sendLowPriorityWork([=]() { performPeriodicChecks(); });
+}
+
 void StreamController::checkIfStreamsAreDone() {
   try {
     Streamers.erase(
         std::remove_if(Streamers.begin(), Streamers.end(),
                        [](auto const &Elem) { return Elem->isDone(); }),
         Streamers.end());
-
     if (Streamers.empty()) {
       StreamersRemaining.store(false);
     }
-    std::this_thread::sleep_for(50ms);
-    Executor.sendLowPriorityWork([=]() { checkIfStreamsAreDone(); });
   } catch (std::exception &E) {
     HasError = true;
     std::lock_guard Guard(ErrorMsgMutex);
@@ -154,6 +179,22 @@ void StreamController::checkIfStreamsAreDone() {
         fmt::format("Got stream error. The error message was: {}", E.what());
     stop();
     StreamersRemaining.store(false);
+  }
+}
+
+void StreamController::throttleIfWriteQueueIsFull() {
+  auto QueuedWrites = WriterThread.nrOfWritesQueued();
+  if (QueuedWrites > StreamerOptions.MaxQueuedWrites &&
+      !StreamersPaused.load()) {
+    LOG_DEBUG("Maximum queued writes exceeded (count={}). Pausing consumers...",
+              QueuedWrites);
+    pauseStreamers();
+  } else if (QueuedWrites < QueuedWritesResumeThreshold *
+                                StreamerOptions.MaxQueuedWrites &&
+             StreamersPaused.load()) {
+    LOG_DEBUG("Write queue below maximum (count={}). Resuming consumers...",
+              QueuedWrites);
+    resumeStreamers();
   }
 }