+++ src/StreamerOptions.h
@@ -23,6 +23,7 @@ struct StreamerOptions {
   time_point StopTimestamp{time_point::max()};
   duration BeforeStartTime{10s};
   duration AfterStopTime{10s};
+  size_t MaxQueuedWrites{1000};
 };
 
 } // namespace FileWriter