+++ changes.md
@@ -2,6 +2,9 @@
 
 ## Next version
 
+- Feat: Cap the size of the output buffer (`max-queued-writes` setting) to
+  avoid out-of-memory issues. Kafka consumers are paused until the buffer
+  usage falls below a fixed threshold.
 - Adding (tcp based) service api to query a filewriter for its status
 - Fix: Cannot import extra modules due to mismatch in ep00, ep01 and al00 registered names